asdf

+++ day11/src/oop/method4/Menu.java
@@ -12,7 +12,7 @@ public class Menu {
 		void infor() {
 			if(dis != 0) {
 				System.out.println("행사중"+this.name +"\t"+ this.catego );
-				System.out.println(this.price - (this.price *(dis/100d)));
+				System.out.println((int) (this.price - (this.price *(dis/100d))));
 				System.out.println();
 			}else {
 				System.out.println(this.name +"\t"+ this.catego );
